requirements.txt
python-telegram-bot==20.7
streamlit
pandas
plotly
streamlit-authenticator
requests